Finished implementing camera tracking and rotation
