--- Assets/_Scripts/Camera/CameraFollow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 821d63a3dbb10a842b39fd82cbb8f713
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
